dedupe luci_common.dart code
