+++ lib/web_ui/dev/firefox_installer.dart
@@ -11,7 +11,7 @@ import 'package:path/path.dart' as path;
 import 'package:yaml/yaml.dart';
 
 import 'common.dart';
-import 'environment.dart';
+import 'package:luci/luci_common.dart';
 import 'exceptions.dart';
 
 class FirefoxArgParser extends BrowserArgParser {
