+++ lib/web_ui/dev/licenses.dart
@@ -8,7 +8,7 @@ import 'dart:io' as io;
 import 'package:args/command_runner.dart';
 import 'package:path/path.dart' as path;
 
-import 'environment.dart';
+import 'package:luci/luci_common.dart';
 
 class LicensesCommand extends Command<bool> {
   @override
